Fixed combined states
- states are now combined using bitflags.
-50000+ episodes trained.

+++ Pacman-Clone-master/Src/Audio/AudioManager.h
@@ -9,7 +9,7 @@ class AudioManager
 {
 public:
 	AudioManager();
-	void PlaySound(Sounds soundType, bool loop, int volume);
+	void PlaySound(Sounds soundType, bool loop, float volume);
 	void StopSound(Sounds soundType = Sounds::None);
 	bool IsPlayingAudio(Sounds soundType);
 	void SetVolume(float volume);
